tweaks

+++ src/Success.js
@@ -6,15 +6,17 @@ import styles from './Success.scss';
 
 class Success extends React.Component {
   componentDidMount() {
-    const user = firebase.auth().currentUser;
-    user.delete();
+    const user = Firebase.auth().currentUser;
+    if (user) {
+      user.delete();
+    }
   }
 
   render() {
     return (
       <div className={styles.container}>
         <h1>Success!</h1>
-        <p>Your Totobee has been logged successfully!</p>
+        <p>Your Totobee visit has been logged successfully!</p>
         <p>
           Now, put your Totobee somewhere so another lucky person can find it!
         </p>
